// graphql_express/src/utils/errors/error-types.ts
export enum AUTH_ERROR_TYPES {
  AUTH_NO_TOKEN = "AUTH_NO_TOKEN",
  AUTH_PASSWORD = "AUTH_PASSWORD",
  AUTH_NOT_LOGIN = "AUTH_NOT_LOGIN",
  AUTH_NOTFOUND = "AUTH_NOTFOUND",
  AUTH_EXPIRED = "AUTH_EXPIRED",
  AUTH_TOKEN = "AUTH_TOKEN",
  AUTH_NOTAUTHORIZED = "AUTH_NOTAUTHORIZED"
}

export enum DATABASE_ERROR_TYPES {
  DATABASE_UNKNOWN = "DATABASE_UNKNOWN",
  DATABASE_NOTFOUND = "DATABASE_NOTFOUND"
}

export enum GENERAL_ERROR_TYPES {
  GENERAL_INPUT = "GENERAL_INPUT",
  GENERAL_FILETYPE = "GENERAL_FILETYPE",
  GENERAL_FILE_SIZE = "GENERAL_FILE_SIZE",
  GENERAL_FILE_MOVE = "GENERAL_FILE_MOVE",
  GENERAL_FILE_REMOVE = "GENERAL_FILE_REMOVE",
  GENERAL_FILE_NOTFOUND = "GENERAL_FILE_NOTFOUND",
  GENERAL_QUERY = "GENERAL_QUERY",
  GENERAL_UNASSIGNABLE = "GENERAL_UNASSIGNABLE",
  GENERAL_DUPLICATE = "GENERAL_DUPLICATE"
}

export enum USERS_ERROR_TYPES {
  USERS = "USERS",
  LEVEL = "LEVEL"
}

export enum FILE_ERRORS {
  FILE_SIZE_ERROR = "FILE_SIZE_ERROR",
  FILE_TYPE_ERROR = "FILE_TYPE_ERROR",
  FILE_MOVE_ERROR = "FILE_MOVE_ERROR",
  FILE_READ_ERROR = "FILE_READ_ERROR",
  FILE_DELETE_ERROR = "FILE_DELETE_ERROR"
}

export enum EXCEL_ERRORS {
  INVALID_DATE_ERROR = "INVALID_DATE_ERROR",
  INVALID_FIELD_ERROR = "INVALID_FIELD_ERROR"
}

export enum INVOICE_ERROR_TYPES {
  INVOICE_INVALID_TOTAL_AMOUNT = "INVOICE_INVALID_TOTAL_AMOUNT",
  INVOICE_INVALID_STATUS = "INVOICE_INVALID_STATUS",
  INVOICE_COUNT_LIMIT = "INVOICE_COUNT_LIMIT",
  INVOICE_CANCELED_DUPLICATE = "INVOICE_CANCELED_DUPLICATE",
  INVOICE_CANCELED_INVALID = "INVOICE_CANCELED_INVALID"
}